Modified the file to pull from Win32::Registry
The previous version, pulling from WMI, was incredibly slow, ranging from 20s on one of my faster machines to approx. 6 minutes on one of my slower ones. I have yet to have this result in a time over 2s.

The downside to this version is that it is noticeably messier, the formatting does not look as nice and some files do not seem to have all attributes stored.